requirements.txt
asgiref
autopep8==1.5.3
bson==0.5.8
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.13
django-ajax-selects==1.9.1
django-cors-headers==3.3.0
djangorestframework==3.11.0
djongo==1.3.1
gunicorn==20.0.4
Pillow==7.1.2
pycodestyle==2.6.0
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
simplejson==3.17.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
Unipath==1.1
whitenoise==5.1.0
